+++ frontend/tests/edit-app-testutils.js
@@ -1,6 +1,7 @@
 import reactTestUtils, {blockUtils} from './my-test-utils.js';
 import store, {observeStore, setCurrentPage, selectOpQueue} from '../edit-app/src/store.js';
 import AddPageMainPanelView from '../edit-app/src/AddPageMainPanelView.jsx';
+import CreatePageTypeMainPanelView from '../edit-app/src/CreatePageTypeMainPanelView.jsx';
 import BlockTrees from '../edit-app/src/BlockTrees.jsx';
 import blockTypes from '../edit-app/src/block-types/block-types.js';
 import InspectorPanel from '../edit-app/src/InspectorPanel.jsx';
@@ -8,20 +9,29 @@ import SaveButton from '../edit-app/src/SaveButton.jsx';
 import blockTreeUtils from '../edit-app/src/blockTreeUtils.js';
 import EditAppAwareWebPage from '../webpage/src/EditAppAwareWebPage.js';
 
-const mockPageTypes = [{"name": "Pages", "ownFields": [],}];
+const mockPageTypes = [{
+    name: 'Pages',
+    defaultLayoutId: '1',
+    isListable: true,
+    ownFields: [{}],
+}];
+const draftPageType = Object.assign({}, mockPageTypes[0], {
+    name: 'Draft',
+    defaultFields: {title: {defaultValue: 'Title'}},
+});
 
 const testBlocks = [{
-    "type": "Paragraph",
-    "title": "",
-    "renderer": "sivujetti:block-auto",
-    "id": 'EsmAW0--AnViMcwnIaDP',
-    "children": [],
-    "propsData": [
-        {"key": "text", "value": "© Mysite"},
-        {"key": "cssClass", "value": ""}
+    type: 'Paragraph',
+    title: '',
+    renderer: 'sivujetti:block-auto',
+    id: 'EsmAW0--AnViMcwnIaDP',
+    children: [],
+    propsData: [
+        {key: 'text', value: '© Mysite'},
+        {key: 'cssClass', value: ''}
     ],
-    "text": "© Mysite",
-    "cssClass": ""
+    text: '© Mysite',
+    cssClass: ''
 }];
 
 function createTestState() {
@@ -32,129 +42,129 @@ function createTestState() {
 
 class MockEditApp extends preact.Component {
     /**
-     * @param {{view: 'DefaultView'|'AddPageView'; pref: preact.Ref; onPageLoaded?: () => void;}} props
+     * @param {{view: 'DefaultView'|'AddPageView'|'CreatePageTypeView'; pref: preact.Ref;}} props
      */
-    render({view, pref, onPageLoaded}) {
+    render({view, pref}) {
+        let viewContent = null;
+        if (view === 'AddPageView') viewContent = <AddPageMainPanelView
+            pageType={ mockPageTypes[0] }
+            ref={ pref }
+            blockTreesRef={ preact.createRef() }
+            getLayouts={ () => Promise.resolve([{id: '1', friendlyName: 'Default'}]) }
+            initialLayoutId='1'
+            noAutoFocus/>;
+        else if (view === 'CreatePageTypeView') viewContent = <CreatePageTypeMainPanelView
+            pageType={ draftPageType }
+            ref={ pref }
+            blockTreesRef={ preact.createRef() }
+            getLayouts={ () => Promise.resolve([{id: '1', friendlyName: 'Default'}]) }
+            onPageTypeCreated={ () => {} }/>;
+        else viewContent = <BlockTrees
+            containingView='DefaultMainPanelView'
+            ref={ pref }/>;
         return <>
             <SaveButton/>
-            { view === 'DefaultView'
-                ? <BlockTrees
-                    containingView="DefaultMainPanelView"
-                    onWebPageLoadHandled={ onPageLoaded }
-                    ref={ pref }/>
-                : <AddPageMainPanelView
-                    pageType={ mockPageTypes[0] }
-                    ref={ pref }
-                    blockTreesRef={ preact.createRef() }
-                    getLayouts={ () => Promise.resolve([{id: '1', friendlyName: 'Default'}]) }
-                    initialLayoutId="1"
-                    noAutoFocus/>
-            }
+            { viewContent }
             <InspectorPanel rootEl={ document.getElementById('render-container-el') } outerEl={ document.getElementById('render-container-el') }/>
         </>;
     }
 }
 
-function renderMockEditAppIntoDom(view, then = null) {
+function renderMockEditAppIntoDom(view, onCmpReferenced = _cmp => Promise.resolve()) {
     return new Promise(resolve => {
-        if (view === 'DefaultView')
+        if (view === 'DefaultView' || view === 'AddPageView' || view === 'CreatePageTypeView')
             reactTestUtils.renderIntoDocument(MockEditApp, {
                 view,
-                pref: cmp => { then(cmp); },
-                onPageLoaded: () => { resolve(); }
-            });
-        else if (view === 'AddPageView')
-            reactTestUtils.renderIntoDocument(MockEditApp, {
-                view,
-                pageType: mockPageTypes[0],
                 pref: cmp => {
-                    if (cmp) resolve(cmp);
-                }
+                    if (cmp) onCmpReferenced(cmp).then(() => {
+                        resolve(cmp);
+                    });
+                },
             });
         else
-            throw new Error(`View (${view}) must be 'DefaultView', or 'AddPageView'`);
+            throw new Error(`View (${view}) must be 'DefaultView', 'AddPageView', or 'CreatePageTypeView'`);
     });
 }
 
-function simulatePageLoad(_s, isNewPage = false, testBlocksBundle = 'default') {
+function simulatePageLoad(_s, isNewPage = false, testBlocksBundle = 'default', pageType = 'Pages') {
     const pageBlocks = [{
-        "type": "PageInfo",
-        "title": "",
-        "renderer": "sivujetti:block-auto",
-        "id": "gBPJW0--vD-s3HsG3cu-",
-        "propsData": [{"key": "overrides", "value": "[]"}],
-        "children": [],
-        "overrides": "[]",
-        "parentBlockIdPath": "",
-        "isStoredTo": "page"
+        type: 'PageInfo',
+        title: '',
+        renderer: 'sivujetti:block-auto',
+        id: 'gBPJW0--vD-s3HsG3cu-',
+        propsData: [{key: 'overrides', value: '[]'}],
+        children: [],
+        overrides: '[]',
+        parentBlockIdPath: '',
+        isStoredTo: 'page'
     }, {
-        "type": "Section",
-        "title": "",
-        "renderer": "sivujetti:block-generic-wrapper",
-        "id": '-MfgGtK5pnuk1s0Kws4u',
-        "propsData": [
-            {"key": "bgImage", "value": ""},
-            {"key": "cssClass", "value": "initial-section"}
+        type: 'Section',
+        title: '',
+        renderer: 'sivujetti:block-generic-wrapper',
+        id: '-MfgGtK5pnuk1s0Kws4u',
+        propsData: [
+            {key: 'bgImage', value: ''},
+            {key: 'cssClass', value: 'initial-section'}
         ],
-        "bgImage": "",
-        "cssClass": "initial-section",
-        "children": [{
-            "type": "Heading",
-            "title": "",
-            "renderer": "sivujetti:block-auto",
-            "id": '-Me3jYWcEOlLTgJhzqL8',
-            "children": testBlocksBundle.startsWith('withNestedBlock') ? [{
-                "type": "Paragraph",
-                "title": "",
-                "renderer": "sivujetti:block-auto",
-                "id": '-MlOaY1tZtpkrwPbyuTW',
-                "children": [],
-                "propsData": [
-                    {"key": "text", "value": "Sub text"},
-                    {"key": "level", "value": 2},
-                    {"key": "cssClass", "value": "subtitle"}
+        bgImage: '',
+        cssClass: 'initial-section',
+        children: [{
+            type: 'Heading',
+            title: '',
+            renderer: 'sivujetti:block-auto',
+            id: '-Me3jYWcEOlLTgJhzqL8',
+            children: testBlocksBundle.startsWith('withNestedBlock') ? [{
+                type: 'Paragraph',
+                title: '',
+                renderer: 'sivujetti:block-auto',
+                id: '-MlOaY1tZtpkrwPbyuTW',
+                children: [],
+                propsData: [
+                    {key: 'text', value: 'Sub text'},
+                    {key: 'level', value: 2},
+                    {key: 'cssClass', value: 'subtitle'}
                 ],
-                "text": "Sub text",
-                "level": 2,
-                "cssClass": "subtitle"
+                text: 'Sub text',
+                level: 2,
+                cssClass: 'subtitle'
             }] : [],
-            "propsData": [
-                {"key": "text", "value": "My page"},
-                {"key": "level", "value": 2},
-                {"key": "cssClass", "value": ""}
+            propsData: [
+                {key: 'text', value: 'My page'},
+                {key: 'level', value: 2},
+                {key: 'cssClass', value: ''}
             ],
-            "text": "My page",
-            "level": 2,
-            "cssClass": ""
+            text: 'My page',
+            level: 2,
+            cssClass: ''
         }, {
-            "type": "Paragraph",
-            "title": "",
-            "renderer": "sivujetti:block-auto",
-            "id": '-Me3jYWcEOlLTgJhzqLA',
-            "children": [],
-            "propsData": [
-                {"key": "text", "value": "Hello"},
-                {"key": "cssClass", "value": ""}
+            type: 'Paragraph',
+            title: '',
+            renderer: 'sivujetti:block-auto',
+            id: '-Me3jYWcEOlLTgJhzqLA',
+            children: [],
+            propsData: [
+                {key: 'text', value: 'Hello'},
+                {key: 'cssClass', value: ''}
             ],
-            "text": "Hello",
-            "cssClass": ""
+            text: 'Hello',
+            cssClass: ''
         }]
     }].concat(...(testBlocksBundle === 'withGlobalBlockReference' ? [{
-        "type": "GlobalBlockReference",
-        "title": "",
-        "renderer": "sivujetti:block-auto",
-        "id": "-MlPupIvfXCsi7eHeTP4",
-        "propsData": [{"key": "globalBlockTreeId", "value": "10"},
-                      {"key": "overrides", "value": "{}"},
-                      {"key": "useOverrides", "value": 0}],
-        "children": [],
-        "globalBlockTreeId": "10",
-        "overrides": "{}",
-        "useOverrides": 0,
-        "__globalBlockTree": {
-            "id": "10",
-            "name": "My stored",
-            "blocks": testBlocks.slice(0),
+        type: 'GlobalBlockReference',
+        title: '',
+        renderer: 'sivujetti:block-auto',
+        id: '-MlPupIvfXCsi7eHeTP4',
+        propsData: [{key: 'globalBlockTreeId', value: '10'},
+                      {key: 'overrides', value: '{}'},
+                      {key: 'useOverrides', value: 0}],
+        children: [],
+        globalBlockTreeId: '10',
+        overrides: '{}',
+        useOverrides: 0,
+        __globalBlockTree: {
+            id: '10',
+            name: 'My stored',
+            blocks: testBlocks.slice(0),
         },
     }] : []));
     //
@@ -186,7 +196,7 @@ function simulatePageLoad(_s, isNewPage = false, testBlocksBundle = 'default') {
             id: '1',
             title: 'Page',
             isPlaceholderPage: isNewPage,
-            type: 'Pages',
+            type: pageType,
             layoutId: '1',
             blocks: pageBlocks,
         },
@@ -197,6 +207,7 @@ function simulatePageLoad(_s, isNewPage = false, testBlocksBundle = 'default') {
     const combinedBlockTree = webPage.getCombinedAndOrderedBlockTree(pageBlocks,
         blockRefs, blockTreeUtils);
     store.dispatch(setCurrentPage({webPage, combinedBlockTree, blockRefs}));
+    return Promise.resolve();
 }
 
 function waitUntiSaveButtonHasRunQueuedOps() {