+++ frontend/tests/utils/create-stuff-test-utils.js
@@ -1,0 +1,32 @@
+import {observeStore, selectCurrentPage} from '../../edit-app/src/store.js';
+import * as appTestUtils from '../edit-app-testutils.js';
+
+/**
+ * @param {Boolean} isDrafPageType = false
+ * @returns {Promise<void>}
+ */
+function simulatePlaceholderPageLoad(isDrafPageType = false) {
+    // Wait for next "selectCurrentPage"
+    const p = new Promise(resolve => {
+        observeStore(selectCurrentPage, () => {
+            resolve();
+        });
+    });
+    // Simulate page load (which triggers "selectCurrentPage")
+    appTestUtils.simulatePageLoad(null, true, ...(!isDrafPageType ? [] : [undefined, 'Draft']));
+    return p;
+}
+
+/**
+ * @param {any} _s
+ * @returns {Promise<void>}
+ */
+function clickSubmitButton(_s) {
+    const p = appTestUtils.waitUntiSaveButtonHasRunQueuedOps();
+    setTimeout(() => {
+        document.querySelector('#render-container-el button').click();
+    }, 1);
+    return p;
+}
+
+export {simulatePlaceholderPageLoad, clickSubmitButton};