+++ changelog.md
@@ -5,6 +5,7 @@
 # 0.8.0-dev (not yet released)
 
 ## Added
+- An ability to create new page types (from the edip app's main menu)
 - tagType -property to Button blocks (`link` -> `<a href ...>`, `button` -> `<button type="button" ...>`, `submit` -> `<button type="submit" ...>`)
 - `UserSite|PluginApi->getPlugin(string $name)` so plugins can publish their own APIs.